# backend/openapi/schema.yml
openapi: 3.0.3
info:
  title: SOBASED API
  version: 1.0.0
paths:
...
  /api/transactions/:
    get:
      operationId: transactions_list
      description: |-
        ProjectScopedViewSet whose custom GET actions (receipt...) only need
        the reading side of the finance rights.
      parameters:
      - in: query
        name: category
        schema:
          type: integer
      - in: query
        name: date_after
        schema:
          type: string
          format: date
      - in: query
        name: date_before
        schema:
          type: string
          format: date
      - in: query
        name: event
        schema:
          type: integer
      - in: query
        name: include_descendants
        schema:
          type: boolean
      - in: query
        name: kind
        schema:
          type: string
          enum:
          - expense
          - income
      - in: query
        name: needs_receipt
        schema:
          type: boolean
      - name: ordering
        required: false
        in: query
        description: Quel champ utiliser pour classer les résultats.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Un numéro de page de l'ensemble des résultats.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Nombre de résultats à retourner par page.
        schema:
          type: integer
      - in: query
        name: paid_by
        schema:
          type: string
        description: '`me` ou nom d''utilisateur'
      - in: query
        name: paid_by_contact
        schema:
          type: integer
      - in: query
        name: payment_status
        schema:
          type: string
          enum:
          - paid
          - to_pay
      - in: query
        name: project
        schema:
          type: integer
      - in: query
        name: reimbursed
        schema:
          type: boolean
      - in: query
        name: search
        schema:
          type: string
      - in: query
        name: to_reimburse
        schema:
          type: boolean
        description: Avances ouvertes
      - in: query
        name: workspace
        schema:
          type: integer
      tags:
      - transactions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionList'
          description: ''
    post:
      operationId: transactions_create
      description: |-
        ProjectScopedViewSet whose custom GET actions (receipt...) only need
        the reading side of the finance rights.
      tags:
      - transactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
          description: ''
...
components:
  schemas:
...
    DisplayStatusEnum:
      enum:
      - to_pay
      - needs_receipt
      - to_reimburse
      - ok
      type: string
      description: |-
        * `to_pay` - À payer
        * `needs_receipt` - À justifier
        * `to_reimburse` - À rembourser
        * `ok` - OK
...
    PaginatedTransactionList:
      type: object
      required:
      - count
      - next
      - previous
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...
    Payer:
      type: object
      description: 'Who advanced the money: a user (by username) or a contact.'
      properties:
        type:
          $ref: '#/components/schemas/PayerTypeEnum'
        id:
          type: integer
          description: Contact id (users have none)
        username:
          type: string
          nullable: true
        name:
          type: string
      required:
      - id
      - name
      - type
      - username
    PayerTypeEnum:
      enum:
      - user
      - contact
      - unknown
      type: string
      description: |-
        * `user` - user
        * `contact` - contact
        * `unknown` - unknown
    PaymentStatusEnum:
      enum:
      - to_pay
      - paid
      type: string
      description: |-
        * `to_pay` - À payer
        * `paid` - Payé
...
    PublicUser:
      type: object
      description: |-
        The ONLY shape in which a user is exposed to other users (SPEC §4).

        Never add e-mail, phone or id here: the search endpoint is open to every
        signed-in user.
      properties:
        username:
          type: string
          title: Nom d'utilisateur
          pattern: ^[A-Za-z0-9_.-]{3,30}$
          maxLength: 30
        display_name:
          type: string
          readOnly: true
        avatar_url:
          type: string
          nullable: true
          readOnly: true
      required:
      - avatar_url
      - display_name
      - username
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        project:
          type: integer
        project_name:
          type: string
          readOnly: true
        project_color:
          type: string
          readOnly: true
        kind:
          $ref: '#/components/schemas/TransactionKindEnum'
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        date:
          type: string
          format: date
        category:
          type: integer
        category_name:
          type: string
          readOnly: true
        label:
          type: string
          maxLength: 200
        vendor:
          type: string
          maxLength: 120
        contact:
          type: integer
          nullable: true
        contact_name:
          type: string
          nullable: true
          readOnly: true
        event:
          type: integer
          nullable: true
        event_title:
          type: string
          nullable: true
          readOnly: true
        has_receipt:
          type: boolean
          readOnly: true
        receipt_name:
          type: string
          readOnly: true
        receipt_content_type:
          type: string
          readOnly: true
        receipt_url:
          type: string
          nullable: true
          readOnly: true
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        paid_by_contact:
          type: integer
          nullable: true
        payer:
          allOf:
          - $ref: '#/components/schemas/Payer'
          nullable: true
          readOnly: true
        to_reimburse:
          type: boolean
        reimbursed_on:
          type: string
          format: date
          nullable: true
        recurring_expense:
          type: integer
          readOnly: true
          nullable: true
        period_key:
          type: string
          readOnly: true
        needs_receipt:
          type: boolean
          readOnly: true
        is_to_reimburse:
          type: boolean
          readOnly: true
        display_status:
          allOf:
          - $ref: '#/components/schemas/DisplayStatusEnum'
          readOnly: true
        created_by:
          allOf:
          - $ref: '#/components/schemas/PublicUser'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - category
      - category_name
      - contact
      - contact_name
      - created_at
      - created_by
      - date
      - display_status
      - event
      - event_title
      - has_receipt
      - id
      - is_to_reimburse
      - kind
      - label
      - needs_receipt
      - paid_by_contact
      - payer
      - payment_status
      - period_key
      - project
      - project_color
      - project_name
      - receipt_content_type
      - receipt_name
      - receipt_url
      - recurring_expense
      - reimbursed_on
      - to_reimburse
      - updated_at
      - vendor
    TransactionKindEnum:
      enum:
      - expense
      - income
      type: string
      description: |-
        * `expense` - Dépense
        * `income` - Recette
    TransactionRequest:
      type: object
      properties:
        project:
          type: integer
        kind:
          $ref: '#/components/schemas/TransactionKindEnum'
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        date:
          type: string
          format: date
        category:
          type: integer
        label:
          type: string
          minLength: 1
          maxLength: 200
        vendor:
          type: string
          maxLength: 120
        contact:
          type: integer
          nullable: true
        event:
          type: integer
          nullable: true
        receipt:
          type: string
          format: binary
          writeOnly: true
          nullable: true
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        paid_by_username:
          type: string
          writeOnly: true
          nullable: true
        paid_by_contact:
          type: integer
          nullable: true
        to_reimburse:
          type: boolean
        reimbursed_on:
          type: string
          format: date
          nullable: true
      required:
      - amount
      - category
      - date
      - kind
      - label
      - project
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid